+++ packages/eslint-plugin-esm/src/rules/no-declaration-file-imports.spec.ts
@@ -1,4 +1,4 @@
-import { test } from "../test.spec.js";
+import { test } from "@fenge/dev-utils";
 import { noDeclarationFileImports } from "./no-declaration-file-imports.js";
 
 const invalid = [