+++ packages/eslint-plugin/src/rules/no-instanceof-builtin.spec.ts
@@ -1,4 +1,4 @@
-import { test } from "../test.spec.js";
+import { test } from "@fenge/dev-utils";
 import { noInstanceofBuiltin } from "./no-instanceof-builtin.js";
 
 const invalid = [